+++ lib/view_modifiers/view_modifier.dart
@@ -1,8 +1,12 @@
 
-export 'general_decorations.dart';
-export 'padding_margin.dart';
-export 'positioning.dart';
-export 'gesture_handlers.dart';
-export 'visibility_opacity.dart';
-export 'transformations.dart';
-export 'layout.dart';
+
+export 'general_decorations_extensions.dart';
+export 'padding_margin_extensions.dart';
+export 'positioning_extensions.dart';
+export 'gesture_handlers_extensions.dart';
+export 'visibility_opacity_extensions.dart';
+export 'transformations_extensions.dart';
+export 'layout_extensions.dart';
+export 'num_extensions.dart';
+export 'datetime_extensions.dart';
+export 'string_extensions.dart';