+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.6]
+- Added new **Number Extensions** (int, double) for durations, comparisons, and utility methods.
+- Added new **Duration Extensions** for easier time handling and utility methods.
+- Added new **DateTime Extensions** for date calculations, formatting, and difference calculations.
+- Added new **String Extensions** for validation, formatting, parsing, and utility methods.
+
 ## [0.0.5]
 - Demo Image Added